+++ script.js
@@ -44,6 +44,7 @@ import { init as initShare } from './modules/share/index.js';
 import { init as initAdvancedSearch } from './modules/advanced_search/index.js';
 import { init as initMcpExplorer } from './modules/mcp_explorer/index.js';
 import { init as initGraph } from './modules/graph/index.js';
+import { init as initSelectionActions } from './modules/selection_actions/index.js';
 import { api } from './modules/core/api.js';
 import { state } from './modules/core/state.js';
 
@@ -158,6 +159,7 @@ const init = async () => {
     initShare();
     initAdvancedSearch();
     initMcpExplorer();
+    initSelectionActions();
     initGraph({ onNavigate: navigateToPageId });
     initNav({
         onNavigate: async (id, space, path) => {